+++ src/public/ConfirmedAssigments.ts
@@ -1,8 +1,8 @@
-import { VariantType } from './Experiment';
+import { Variant } from './Experiment';
 
 export class ConfirmedAssignment {
   experimentId: String;
-  variant: VariantType;
+  variant: Variant;
 
   constructor(experimentId: String, variant: VariantType) {
     this.experimentId = experimentId;
@@ -12,7 +12,7 @@ export class ConfirmedAssignment {
   static fromJson(json: any): ConfirmedAssignment {
     return new ConfirmedAssignment(
       json.experimentId,
-      VariantType[json.variant as keyof typeof VariantType]
+      Variant.fromJson(json.variant)
     );
   }
 }